update csv output
